+++ test/Integration/Gateways/GpApiConnector/DccCardNotPresentTest.php
@@ -3,6 +3,7 @@
 use GlobalPayments\Api\Entities\Enums\Environment;
 use GlobalPayments\Api\Entities\Enums\Channel;
 use GlobalPayments\Api\Entities\Enums\TransactionStatus;
+use GlobalPayments\Api\Entities\Exceptions\GatewayException;
 use GlobalPayments\Api\PaymentMethods\CreditCardData;
 use GlobalPayments\Api\ServiceConfigs\Gateways\GpApiConfig;
 use GlobalPayments\Api\ServicesContainer;
@@ -16,7 +17,7 @@ class DccCardNotPresentTest extends TestCase
 {
     private $currency = 'EUR';
     private $amount = 15.11;
-    /** @var CreditCardData  */
+    /** @var CreditCardData */
     private $card;
 
     public function setup()
@@ -49,9 +50,9 @@ class DccCardNotPresentTest extends TestCase
         $orderId = GenerationUtils::generateOrderId();
 
         $dccDetails = $this->card->getDccRate()
-                ->withAmount($this->amount)
-                ->withCurrency($this->currency)
-                ->execute();
+            ->withAmount($this->amount)
+            ->withCurrency($this->currency)
+            ->execute();
 
         $this->assertNotNull($dccDetails);
         $this->assertEquals('SUCCESS', $dccDetails->responseCode);
@@ -265,5 +266,123 @@ class DccCardNotPresentTest extends TestCase
         $this->assertNotNull($response);
         $this->assertEquals('SUCCESS', $response->responseCode);
         $this->assertEquals(TransactionStatus::CAPTURED, $response->responseMessage);
+    }
+
+    public function testCreditGetDccInfo_WithIdempotencyKey()
+    {
+        $idempotency = GenerationUtils::getGuid();
+
+        $dccDetails = $this->card->getDccRate()
+            ->withAmount($this->amount)
+            ->withCurrency($this->currency)
+            ->withIdempotencyKey($idempotency)
+            ->execute();
+
+        $this->assertNotNull($dccDetails);
+        $this->assertEquals('SUCCESS', $dccDetails->responseCode);
+        $this->assertEquals('AVAILABLE', $dccDetails->responseMessage);
+        $this->assertNotNull($dccDetails->dccRateData);
+
+        sleep(2);
+
+        $exceptionCaught = false;
+        try {
+            $this->card->getDccRate()
+                ->withAmount($this->amount)
+                ->withCurrency($this->currency)
+                ->withIdempotencyKey($idempotency)
+                ->execute();
+        } catch (GatewayException $e) {
+            $exceptionCaught = true;
+            $this->assertEquals('40039', $e->responseCode);
+            $this->assertEquals(
+                sprintf("Status Code: %s - Idempotency Key seen before: id=%s, status=%s", 'DUPLICATE_ACTION', $dccDetails->transactionId, 'AVAILABLE'
+                ), $e->getMessage()
+            );
+        } finally {
+            $this->assertTrue($exceptionCaught);
+        }
+    }
+
+    public function testCreditGetDccInfo_RateNotAvailable()
+    {
+        $orderId = GenerationUtils::generateOrderId();
+        $this->card->number = "4263970000005262";
+
+        $dccDetails = $this->card->getDccRate()
+            ->withAmount($this->amount)
+            ->withCurrency($this->currency)
+            ->execute();
+
+        $this->assertNotNull($dccDetails);
+        $this->assertEquals('SUCCESS', $dccDetails->responseCode);
+        $this->assertEquals('NOT_AVAILABLE', $dccDetails->responseMessage);
+        $this->assertNotNull($dccDetails->dccRateData);
+
+        sleep(2);
+
+        $response = $this->card->charge($this->amount)
+            ->withCurrency($this->currency)
+            ->withAllowDuplicates(true)
+            ->withDccRateData($dccDetails->dccRateData)
+            ->withClientTransactionId($orderId)
+            ->execute();
+
+        $this->assertNotNull($response);
+        $this->assertEquals('SUCCESS', $response->responseCode);
+        $this->assertEquals('CAPTURED', $response->responseMessage);
+        $this->assertEquals($this->amount, $response->dccRateData->cardHolderAmount);
+        $this->assertEquals($this->currency, $response->dccRateData->cardHolderCurrency);
+    }
+
+    public function testCreditGetDccInfo_InvalidCardNUmber()
+    {
+        $this->card->number = "4000000000005262";
+
+        $exceptionCaught = false;
+        try {
+            $this->card->getDccRate()
+                ->withAmount($this->amount)
+                ->withCurrency($this->currency)
+                ->execute();
+        } catch (GatewayException $e) {
+            $exceptionCaught = true;
+            $this->assertEquals('40085', $e->responseCode);
+            $this->assertEquals('Status Code: INVALID_REQUEST_DATA - Card number fails Luhn Check', $e->getMessage());
+        } finally {
+            $this->assertTrue($exceptionCaught);
+        }
+    }
+
+    public function testCreditGetDccInfo_WithoutAmount()
+    {
+        $exceptionCaught = false;
+        try {
+            $this->card->getDccRate()
+                ->withCurrency($this->currency)
+                ->execute();
+        } catch (GatewayException $e) {
+            $exceptionCaught = true;
+            $this->assertEquals('40005', $e->responseCode);
+            $this->assertEquals('Status Code: MANDATORY_DATA_MISSING - Request expects the following fields : amount', $e->getMessage());
+        } finally {
+            $this->assertTrue($exceptionCaught);
+        }
+    }
+
+    public function testCreditGetDccInfo_WithoutCurrency()
+    {
+        $exceptionCaught = false;
+        try {
+            $this->card->getDccRate()
+                ->withAmount($this->amount)
+                ->execute();
+        } catch (GatewayException $e) {
+            $exceptionCaught = true;
+            $this->assertEquals('40005', $e->responseCode);
+            $this->assertEquals('Status Code: MANDATORY_DATA_MISSING - Request expects the following fields : currency', $e->getMessage());
+        } finally {
+            $this->assertTrue($exceptionCaught);
+        }
     }
 }