+++ test/Integration/Gateways/GpApiConnector/DccCardPresentTest.php
@@ -1,24 +1,28 @@
 <?php
 
-use GlobalPayments\Api\Entities\Enums\Environment;
 use GlobalPayments\Api\Entities\Enums\Channel;
+use GlobalPayments\Api\Entities\Enums\EntryMethod;
+use GlobalPayments\Api\Entities\Enums\Environment;
 use GlobalPayments\Api\Entities\Enums\TransactionStatus;
+use GlobalPayments\Api\Entities\Exceptions\GatewayException;
+use GlobalPayments\Api\Entities\GpApi\AccessTokenInfo;
 use GlobalPayments\Api\PaymentMethods\CreditCardData;
+use GlobalPayments\Api\PaymentMethods\CreditTrackData;
+use GlobalPayments\Api\PaymentMethods\DebitTrackData;
 use GlobalPayments\Api\ServiceConfigs\Gateways\GpApiConfig;
 use GlobalPayments\Api\Services\GpApiService;
 use GlobalPayments\Api\ServicesContainer;
 use GlobalPayments\Api\Utils\GenerationUtils;
-use PHPUnit\Framework\TestCase;
 use GlobalPayments\Api\Utils\Logging\Logger;
 use GlobalPayments\Api\Utils\Logging\SampleRequestLogger;
-use \GlobalPayments\Api\Entities\GpApi\AccessTokenInfo;
+use PHPUnit\Framework\TestCase;
 
 
 class DccCardPresentTest extends TestCase
 {
     private $currency = 'EUR';
     private $amount = 15.11;
-    /** @var CreditCardData  */
+    /** @var CreditCardData */
     private $card;
     const DCC_RATE_CONFIG = 'dcc_rate';
 
@@ -214,5 +218,65 @@ class DccCardPresentTest extends TestCase
         $this->assertNotNull($reverse);
         $this->assertEquals('SUCCESS', $reverse->responseCode);
         $this->assertEquals(TransactionStatus::CAPTURED, $reverse->responseMessage);
+    }
+
+    public function testCreditGetDccInfo_CreditTrackData()
+    {
+        $creditTrackData = new CreditTrackData();
+        $creditTrackData->setTrackData('%B4012002000060016^VI TEST CREDIT^251210118039000000000396?;4012002000060016=25121011803939600000?');
+        $creditTrackData->entryMethod = EntryMethod::SWIPE;
+        $orderId = GenerationUtils::generateOrderId();
+
+        $dccDetails = $this->getDccDetails();
+        $this->assertNotNull($dccDetails);
+        $this->assertEquals('SUCCESS', $dccDetails->responseCode);
+        $this->assertEquals('NOT_AVAILABLE', $dccDetails->responseMessage);
+        $this->assertNotNull($dccDetails->dccRateData);
+
+        $exceptionCaught = false;
+        try {
+            $creditTrackData->charge($this->amount)
+                ->withCurrency($this->currency)
+                ->withAllowDuplicates(true)
+                ->withDccRateData($dccDetails->dccRateData)
+                ->withClientTransactionId($orderId)
+                ->execute();
+        } catch (GatewayException $e) {
+            $exceptionCaught = true;
+            $this->assertEquals('40211', $e->responseCode);
+            $this->assertEquals('Status Code: MANDATORY_DATA_MISSING - 37,Request expects the following field  payer_amount payer_currency exchange_rate commission_percentage  from the Merchant.', $e->getMessage());
+        } finally {
+            $this->assertTrue($exceptionCaught);
+        }
+    }
+
+    public function testCreditGetDccInfo_DebitTrackData()
+    {
+        $debitCard = new DebitTrackData();
+        $debitCard->value = '%B4012002000060016^VI TEST CREDIT^251210118039000000000396?;4012002000060016=25121011803939600000?';
+        $debitCard->pinBlock = '32539F50C245A6A93D123412324000AA';
+        $debitCard->entryMethod = EntryMethod::SWIPE;
+        $orderId = GenerationUtils::generateOrderId();
+
+        $dccDetails = $this->getDccDetails();
+        $this->assertNotNull($dccDetails);
+        $this->assertEquals('SUCCESS', $dccDetails->responseCode);
+        $this->assertEquals('NOT_AVAILABLE', $dccDetails->responseMessage);
+        $this->assertNotNull($dccDetails->dccRateData);
+
+        $exceptionCaught = false;
+        try {
+            $debitCard->charge($this->amount)
+                ->withCurrency($this->currency)
+                ->withDccRateData($dccDetails->dccRateData)
+                ->withClientTransactionId($orderId)
+                ->execute();
+        } catch (GatewayException $e) {
+            $exceptionCaught = true;
+            $this->assertEquals('50001', $e->responseCode);
+            $this->assertEquals('Status Code: UNAUTHORIZED_DOWNSTREAM - -21,Unauthorized', $e->getMessage());
+        } finally {
+            $this->assertTrue($exceptionCaught);
+        }
     }
 }