+++ test/Integration/Gateways/GpApiConnector/GpApi3DSecureTest.php
@@ -28,6 +28,8 @@ use GlobalPayments\Api\Tests\Integration\Gateways\ThreeDSecureAcsClient;
 use GlobalPayments\Api\Utils\GenerationUtils;
 use GlobalPayments\Api\Entities\Enums\Secure3dStatus;
 use GlobalPayments\Api\Entities\Enums\ManualEntryMethod;
+use GlobalPayments\Api\Utils\Logging\Logger;
+use GlobalPayments\Api\Utils\Logging\SampleRequestLogger;
 use PHPUnit\Framework\TestCase;
 
 class GpApi3DSecureTest extends TestCase
@@ -107,6 +109,7 @@ class GpApi3DSecureTest extends TestCase
         $config->challengeNotificationUrl = 'https://ensi808o85za.x.pipedream.net/';
         $config->methodNotificationUrl = 'https://ensi808o85za.x.pipedream.net/';
         $config->merchantContactUrl = 'https://enp4qhvjseljg.x.pipedream.net/';
+//        $config->requestLogger = new SampleRequestLogger(new Logger("logs"));
 
         return $config;
     }